fix: Fixed crash if metrics were not registered
If you don't call RegisterMetrics and try to make a request to http using tonic, an error will occur:
panic: runtime error: invalid memory address or nil pointer dereference
This is not fatal, but it is not obvious.
Now the log will see that it is necessary to call RegisterMetrics. The error will go away.